@@ -69,7 +69,7 @@ class GitLab {
   ///
   final bool assumeUtf8;
 
-  final GitLabHttpClient _httpClient;
+  final GitLabHttpClient? _httpClient;
 
   static const String apiVersion = 'v4';
 
@@ -77,7 +77,7 @@ class GitLab {
       {this.host = 'gitlab.com', this.scheme = 'https', this.assumeUtf8 = true})
       : _httpClient = new GitLabHttpClient();
 
-  GitLab._test(GitLabHttpClient httpClient, this.token,
+  GitLab._test(GitLabHttpClient? httpClient, this.token,
       {this.host: 'gitlab.com', this.scheme: 'https', this.assumeUtf8 = true})
       : _httpClient = httpClient;
 
@@ -91,13 +91,13 @@ class GitLab {
   @visibleForTesting
   Future<dynamic> request(Uri uri,
       {HttpMethod method: HttpMethod.get,
-      String body,
+      String? body,
       bool asJson: true}) async {
     final headers = <String, String>{'PRIVATE-TOKEN': token};
 
     _log.fine('Making GitLab $method request to $uri.');
 
-    final response = await _httpClient.request(uri, headers, method);
+    final response = await _httpClient!.request(uri, headers, method);
 
     if (!(response.statusCode >= 200 && response.statusCode < 300)) {
       throw new GitLabException(response.statusCode, response.body);
@@ -116,7 +116,7 @@ class GitLab {
   /// This function is used internally to build the URIs for API calls.
   @visibleForTesting
   Uri buildUri(Iterable<String> pathSegments,
-      {Map<String, dynamic> queryParameters, int page, int perPage}) {
+      {Map<String, dynamic>? queryParameters, int? page, int? perPage}) {
     dynamic _addQueryParameter(String key, dynamic value) =>
         (queryParameters ??= new Map<String, dynamic>())[key] = '$value';
 
@@ -125,7 +125,7 @@ class GitLab {
     return new Uri(
         scheme: scheme,
         host: host,
-        pathSegments: ['api', apiVersion]..addAll(pathSegments),
+        pathSegments: <String>['api', apiVersion]..addAll(pathSegments),
         queryParameters: queryParameters);
   }
 }
@@ -143,6 +143,6 @@ class GitLabException implements Exception {
 /// A helper function to get a [GitLab] instance with a [GitLabHttpClient] that
 /// can be mocked.
 @visibleForTesting
-GitLab getTestable(GitLabHttpClient httpClient,
+GitLab getTestable(GitLabHttpClient? httpClient,
         [String token = 'secret-token']) =>
     new GitLab._test(httpClient, token);
