@@ -9,13 +9,13 @@ class IssuesApi {
   Future<Issue> get(int id) async {
     final uri = _project.buildUri(['issues', '$id']);
 
-    final json = await _gitLab.request(uri) as Map;
+    final json = await _gitLab.request(uri) as Map?;
 
     return new Issue.fromJson(json);
   }
 
   Future<List<Issue>> closedByMergeRequest(int mergeRequestIid,
-      {int page, int perPage}) async {
+      {int? page, int? perPage}) async {
     final uri = _project.buildUri(
       ['merge_requests', '$mergeRequestIid', 'closes_issues'],
       page: page,
@@ -28,13 +28,13 @@ class IssuesApi {
   }
 
   Future<List<Issue>> list(
-      {IssueState state,
-      IssueOrderBy orderBy,
-      IssueSort sort,
-      String milestone,
-      List<String> labels,
-      int page,
-      int perPage}) async {
+      {IssueState? state,
+      IssueOrderBy? orderBy,
+      IssueSort? sort,
+      String? milestone,
+      List<String>? labels,
+      int? page,
+      int? perPage}) async {
     final queryParameters = <String, dynamic>{};
 
     if (state != null) queryParameters['state'] = _enumToString(state);
@@ -57,34 +57,34 @@ enum IssueOrderBy { createdAt, updatedAt }
 enum IssueSort { asc, desc }
 
 class Issue {
-  final Map originalJson;
+  final Map? originalJson;
 
   Issue.fromJson(this.originalJson);
 
-  int get projectId => originalJson['project_id'] as int;
-  int get id => originalJson['id'] as int;
-  int get iid => originalJson['iid'] as int;
+  int? get projectId => originalJson!['project_id'] as int?;
+  int? get id => originalJson!['id'] as int?;
+  int? get iid => originalJson!['iid'] as int?;
 
-  String get title => originalJson['title'] as String;
-  String get description => originalJson['description'] as String;
+  String? get title => originalJson!['title'] as String?;
+  String? get description => originalJson!['description'] as String?;
 
-  String get state => originalJson['state'] as String;
-  List<String> get labels => (originalJson['labels'] as List).cast<String>();
-  String get webUrl => originalJson['web_url'] as String;
+  String? get state => originalJson!['state'] as String?;
+  List<String> get labels => (originalJson!['labels'] as List).cast<String>();
+  String? get webUrl => originalJson!['web_url'] as String?;
 
   DateTime get createdAt =>
-      DateTime.parse(originalJson['created_at'] as String);
+      DateTime.parse(originalJson!['created_at'] as String);
   DateTime get updatedAt =>
-      DateTime.parse(originalJson['updated_at'] as String);
+      DateTime.parse(originalJson!['updated_at'] as String);
 
-  bool get subscribed => originalJson['subscribed'] as bool;
-  int get userNotesCount => originalJson['user_notes_count'] as int;
-  DateTime get dueDate => originalJson['due_date'] == null
+  bool? get subscribed => originalJson!['subscribed'] as bool?;
+  int? get userNotesCount => originalJson!['user_notes_count'] as int?;
+  DateTime? get dueDate => originalJson!['due_date'] == null
       ? null
-      : DateTime.parse(originalJson['due_date'] as String);
+      : DateTime.parse(originalJson!['due_date'] as String);
 
-  bool get confidential => originalJson['confidential'] as bool;
-  int get weight => originalJson['weight'] as int;
+  bool? get confidential => originalJson!['confidential'] as bool?;
+  int? get weight => originalJson!['weight'] as int?;
 
   @override
   String toString() => 'Issue id#$id iid#$iid ($title)';
