@@ -11,7 +11,7 @@ class JobsApi {
   Future<Job> get(int id) async {
     final uri = _project.buildUri(['jobs', '$id']);
 
-    final json = await _gitLab.request(uri) as Map;
+    final json = await _gitLab.request(uri) as Map?;
 
     return new Job.fromJson(json);
   }
@@ -19,12 +19,13 @@ class JobsApi {
   Future<Job> cancel(int id) async {
     final uri = _project.buildUri(['jobs', '$id', 'cancel']);
 
-    final json = await _gitLab.request(uri, method: HttpMethod.post) as Map;
+    final json = await _gitLab.request(uri, method: HttpMethod.post) as Map?;
 
     return new Job.fromJson(json);
   }
 
-  Future<List<Job>> list({List<JobScope> scopes, int page, int perPage}) async {
+  Future<List<Job>> list(
+      {List<JobScope>? scopes, int? page, int? perPage}) async {
     final queryParameters = <String, dynamic>{};
 
     if (scopes != null && scopes.isNotEmpty) {
@@ -48,22 +49,22 @@ class JobsApi {
 enum JobScope { created, pending, running, failed, success, canceled, skipped }
 
 class Job {
-  final Map originalJson;
+  final Map? originalJson;
 
   Job.fromJson(this.originalJson);
 
-  int get id => originalJson['id'] as int;
-  String get name => originalJson['name'] as String;
-  String get ref => originalJson['ref'] as String;
-  String get stage => originalJson['stage'] as String;
-  String get status => originalJson['status'] as String;
+  int? get id => originalJson!['id'] as int?;
+  String? get name => originalJson!['name'] as String?;
+  String? get ref => originalJson!['ref'] as String?;
+  String? get stage => originalJson!['stage'] as String?;
+  String? get status => originalJson!['status'] as String?;
   DateTime get startedAt =>
-      DateTime.parse(originalJson['started_at'] as String);
+      DateTime.parse(originalJson!['started_at'] as String);
   DateTime get createdAt =>
-      DateTime.parse(originalJson['created_at'] as String);
+      DateTime.parse(originalJson!['created_at'] as String);
   DateTime get finishedAt =>
-      DateTime.parse(originalJson['finished_at'] as String);
-  Commit get commit => new Commit.fromJson(originalJson['commit'] as Map);
+      DateTime.parse(originalJson!['finished_at'] as String);
+  Commit get commit => new Commit.fromJson(originalJson!['commit'] as Map?);
 
   @override
   String toString() => 'Build id#$id ($name)';
