@@ -11,12 +11,12 @@ class PipelinesApi {
   Future<Pipeline> get(int id) async {
     final uri = _project.buildUri(['pipelines', '$id']);
 
-    final json = await _gitLab.request(uri) as Map;
+    final json = await _gitLab.request(uri) as Map?;
 
     return new Pipeline.fromJson(json);
   }
 
-  Future<List<Pipeline>> list({int page, int perPage}) async {
+  Future<List<Pipeline>> list({int? page, int? perPage}) async {
     final uri = _project.buildUri(['pipelines'], page: page, perPage: perPage);
 
     final jsonList = _responseToList(await _gitLab.request(uri));
@@ -26,14 +26,14 @@ class PipelinesApi {
 }
 
 class Pipeline {
-  final Map originalJson;
+  final Map? originalJson;
 
   Pipeline.fromJson(this.originalJson);
 
-  int get id => originalJson['id'] as int;
-  String get status => originalJson['status'] as String;
-  String get ref => originalJson['ref'] as String;
-  String get sha => originalJson['sha'] as String;
+  int? get id => originalJson!['id'] as int?;
+  String? get status => originalJson!['status'] as String?;
+  String? get ref => originalJson!['ref'] as String?;
+  String? get sha => originalJson!['sha'] as String?;
 
   @override
   String toString() => 'Pipeline id#$id (Ref: $ref)';
