@@ -10,17 +10,17 @@ class SnippetsApi {
   Future<Snippet> get(int id) async {
     final uri = _project.buildUri(['snippets', '$id']);
 
-    final json = await _gitLab.request(uri) as Map;
+    final json = await _gitLab.request(uri) as Map?;
 
     return new Snippet.fromJson(json);
   }
 
-  Future<String> content(int id) async {
+  Future<String?> content(int id) async {
     final uri = _project.buildUri(['snippets', '$id', 'raw']);
-    return await _gitLab.request(uri, asJson: false) as String;
+    return await _gitLab.request(uri, asJson: false) as String?;
   }
 
-  Future<List<Snippet>> list({int page, int perPage}) async {
+  Future<List<Snippet>> list({int? page, int? perPage}) async {
     final uri = _project.buildUri(['snippets'], page: page, perPage: perPage);
 
     final jsonList = _responseToList(await _gitLab.request(uri));
@@ -29,7 +29,10 @@ class SnippetsApi {
   }
 
   Future update(int id,
-      {String title, String fileName, String code, String visibility}) async {
+      {String? title,
+      String? fileName,
+      String? code,
+      String? visibility}) async {
     final queryParameters = <String, dynamic>{};
 
     if (title != null) queryParameters['title'] = title;
@@ -40,21 +43,21 @@ class SnippetsApi {
     final uri = _project
         .buildUri(['snippets', '$id'], queryParameters: queryParameters);
 
-    final json = await _gitLab.request(uri, method: HttpMethod.put) as Map;
+    final json = await _gitLab.request(uri, method: HttpMethod.put) as Map?;
 
     return new Snippet.fromJson(json);
   }
 }
 
 class Snippet {
-  final Map originalJson;
+  final Map? originalJson;
 
   Snippet.fromJson(this.originalJson);
 
-  int get id => originalJson['id'] as int;
-  String get title => originalJson['title'] as String;
-  String get fileName => originalJson['file_name'] as String;
-  String get webUrl => originalJson['web_url'] as String;
+  int? get id => originalJson!['id'] as int?;
+  String? get title => originalJson!['title'] as String?;
+  String? get fileName => originalJson!['file_name'] as String?;
+  String? get webUrl => originalJson!['web_url'] as String?;
 
   @override
   String toString() => 'Snippet id#$id ($title)';
